Update gfbrowser.cpp
Add optional code to turn on a web inspector window.  Turning the feature on can only be done for operating systems with a window manager.  For the target device this is impossible because there is no window manager, only a single window is displayed.
To activate the optional code, turn it on in gfbrowser.pro by adding DEFINES += DEV_INTERFACE_REQUIRES_MULTI_WINDOW.

Also added QWebSettings::AcceleratedCompositingEnabled, documentation already says this is truned on by default, but the current rendering is pretty slow ~5 seconds to rotate an image.